Commented out Facebook social link.
Facebook keeps asking me to upload a picture of myself in order to login. I did
this once, was verified, and was able to log back in, but within minutes I was
logged out and asked to upload all over again. Too much of a pain; I'll use
other social media.